+++ sdk/resourcemanager/storage/armstorage/deletedaccounts_client.go
@@ -57,7 +57,7 @@ func NewDeletedAccountsClient(subscriptionID string, credential azcore.TokenCred
 
 // Get - Get properties of specified deleted account resource.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-05-01
+// Generated from API version 2022-09-01
 // deletedAccountName - Name of the deleted storage account.
 // location - The location of the deleted storage account.
 // options - DeletedAccountsClientGetOptions contains the optional parameters for the DeletedAccountsClient.Get method.
@@ -96,7 +96,7 @@ func (client *DeletedAccountsClient) getCreateRequest(ctx context.Context, delet
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-05-01")
+	reqQP.Set("api-version", "2022-09-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -112,8 +112,7 @@ func (client *DeletedAccountsClient) getHandleResponse(resp *http.Response) (Del
 }
 
 // NewListPager - Lists deleted accounts under the subscription.
-// If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-05-01
+// Generated from API version 2022-09-01
 // options - DeletedAccountsClientListOptions contains the optional parameters for the DeletedAccountsClient.List method.
 func (client *DeletedAccountsClient) NewListPager(options *DeletedAccountsClientListOptions) *runtime.Pager[DeletedAccountsClientListResponse] {
 	return runtime.NewPager(runtime.PagingHandler[DeletedAccountsClientListResponse]{
@@ -155,7 +154,7 @@ func (client *DeletedAccountsClient) listCreateRequest(ctx context.Context, opti
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-05-01")
+	reqQP.Set("api-version", "2022-09-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil