+++ sdk/resourcemanager/storage/armstorage/managementpolicies_client_example_test.go
@@ -14,10 +14,10 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/storage/armstorage"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/storage/armstorage/v2"
 )
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-05-01/examples/StorageAccountGetManagementPolicy.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountGetManagementPolicy.json
 func ExampleManagementPoliciesClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -36,8 +36,8 @@ func ExampleManagementPoliciesClient_Get() {
 	_ = res
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-05-01/examples/StorageAccountSetManagementPolicy.json
-func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicy() {
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountSetManagementPolicy.json
+func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicies() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -129,7 +129,77 @@ func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementP
 	_ = res
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-05-01/examples/StorageAccountSetManagementPolicyForBlockAndAppendBlobs.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountSetManagementPolicyColdTierActions.json
+func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicyColdTierActions() {
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	if err != nil {
+		log.Fatalf("failed to obtain a credential: %v", err)
+	}
+	ctx := context.Background()
+	client, err := armstorage.NewManagementPoliciesClient("{subscription-id}", cred, nil)
+	if err != nil {
+		log.Fatalf("failed to create client: %v", err)
+	}
+	res, err := client.CreateOrUpdate(ctx, "res7687", "sto9699", armstorage.ManagementPolicyNameDefault, armstorage.ManagementPolicy{
+		Properties: &armstorage.ManagementPolicyProperties{
+			Policy: &armstorage.ManagementPolicySchema{
+				Rules: []*armstorage.ManagementPolicyRule{
+					{
+						Name: to.Ptr("olcmtest1"),
+						Type: to.Ptr(armstorage.RuleTypeLifecycle),
+						Definition: &armstorage.ManagementPolicyDefinition{
+							Actions: &armstorage.ManagementPolicyAction{
+								BaseBlob: &armstorage.ManagementPolicyBaseBlob{
+									Delete: &armstorage.DateAfterModification{
+										DaysAfterModificationGreaterThan: to.Ptr[float32](1000),
+									},
+									TierToArchive: &armstorage.DateAfterModification{
+										DaysAfterModificationGreaterThan: to.Ptr[float32](90),
+									},
+									TierToCold: &armstorage.DateAfterModification{
+										DaysAfterModificationGreaterThan: to.Ptr[float32](30),
+									},
+									TierToCool: &armstorage.DateAfterModification{
+										DaysAfterModificationGreaterThan: to.Ptr[float32](30),
+									},
+								},
+								Snapshot: &armstorage.ManagementPolicySnapShot{
+									Delete: &armstorage.DateAfterCreation{
+										DaysAfterCreationGreaterThan: to.Ptr[float32](30),
+									},
+									TierToCold: &armstorage.DateAfterCreation{
+										DaysAfterCreationGreaterThan: to.Ptr[float32](30),
+									},
+								},
+								Version: &armstorage.ManagementPolicyVersion{
+									Delete: &armstorage.DateAfterCreation{
+										DaysAfterCreationGreaterThan: to.Ptr[float32](30),
+									},
+									TierToCold: &armstorage.DateAfterCreation{
+										DaysAfterCreationGreaterThan: to.Ptr[float32](30),
+									},
+								},
+							},
+							Filters: &armstorage.ManagementPolicyFilter{
+								BlobTypes: []*string{
+									to.Ptr("blockBlob")},
+								PrefixMatch: []*string{
+									to.Ptr("olcmtestcontainer1")},
+							},
+						},
+						Enabled: to.Ptr(true),
+					}},
+			},
+		},
+	}, nil)
+	if err != nil {
+		log.Fatalf("failed to finish the request: %v", err)
+	}
+	// TODO: use response item
+	_ = res
+}
+
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountSetManagementPolicyForBlockAndAppendBlobs.json
 func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicyForBlockAndAppendBlobs() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -185,7 +255,62 @@ func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementP
 	_ = res
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-05-01/examples/StorageAccountSetManagementPolicyWithSnapshotAndVersion.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountSetManagementPolicyHotTierActions.json
+func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicyHotTierActions() {
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	if err != nil {
+		log.Fatalf("failed to obtain a credential: %v", err)
+	}
+	ctx := context.Background()
+	client, err := armstorage.NewManagementPoliciesClient("{subscription-id}", cred, nil)
+	if err != nil {
+		log.Fatalf("failed to create client: %v", err)
+	}
+	res, err := client.CreateOrUpdate(ctx, "res7687", "sto9699", armstorage.ManagementPolicyNameDefault, armstorage.ManagementPolicy{
+		Properties: &armstorage.ManagementPolicyProperties{
+			Policy: &armstorage.ManagementPolicySchema{
+				Rules: []*armstorage.ManagementPolicyRule{
+					{
+						Name: to.Ptr("olcmtest1"),
+						Type: to.Ptr(armstorage.RuleTypeLifecycle),
+						Definition: &armstorage.ManagementPolicyDefinition{
+							Actions: &armstorage.ManagementPolicyAction{
+								BaseBlob: &armstorage.ManagementPolicyBaseBlob{
+									TierToHot: &armstorage.DateAfterModification{
+										DaysAfterModificationGreaterThan: to.Ptr[float32](30),
+									},
+								},
+								Snapshot: &armstorage.ManagementPolicySnapShot{
+									TierToHot: &armstorage.DateAfterCreation{
+										DaysAfterCreationGreaterThan: to.Ptr[float32](30),
+									},
+								},
+								Version: &armstorage.ManagementPolicyVersion{
+									TierToHot: &armstorage.DateAfterCreation{
+										DaysAfterCreationGreaterThan: to.Ptr[float32](30),
+									},
+								},
+							},
+							Filters: &armstorage.ManagementPolicyFilter{
+								BlobTypes: []*string{
+									to.Ptr("blockBlob")},
+								PrefixMatch: []*string{
+									to.Ptr("olcmtestcontainer1")},
+							},
+						},
+						Enabled: to.Ptr(true),
+					}},
+			},
+		},
+	}, nil)
+	if err != nil {
+		log.Fatalf("failed to finish the request: %v", err)
+	}
+	// TODO: use response item
+	_ = res
+}
+
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountSetManagementPolicyWithSnapshotAndVersion.json
 func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicyWithSnapshotAndVersion() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -258,8 +383,8 @@ func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementP
 	_ = res
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-05-01/examples/StorageAccountSetManagementPolicy_BaseBlobDaysAfterCreationActions.json
-func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicy_BaseBlobDaysAfterCreationActions() {
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountSetManagementPolicy_BaseBlobDaysAfterCreationActions.json
+func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicyBaseBlobDaysAfterCreationActions() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -309,8 +434,8 @@ func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementP
 	_ = res
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-05-01/examples/StorageAccountSetManagementPolicy_LastAccessTimeBasedBlobActions.json
-func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicy_LastAccessTimeBasedBlobActions() {
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountSetManagementPolicy_LastAccessTimeBasedBlobActions.json
+func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicyLastAccessTimeBasedBlobActions() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -366,8 +491,8 @@ func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementP
 	_ = res
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-05-01/examples/StorageAccountSetManagementPolicy_LastTierChangeTimeActions.json
-func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicy_LastTierChangeTimeActions() {
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountSetManagementPolicy_LastTierChangeTimeActions.json
+func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementPolicyLastTierChangeTimeActions() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -430,7 +555,7 @@ func ExampleManagementPoliciesClient_CreateOrUpdate_storageAccountSetManagementP
 	_ = res
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-05-01/examples/StorageAccountDeleteManagementPolicy.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/storage/resource-manager/Microsoft.Storage/stable/2022-09-01/examples/StorageAccountDeleteManagementPolicy.json
 func ExampleManagementPoliciesClient_Delete() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {